# server/docs/swagger.yaml
basePath: /ttms
definitions:
  common.State:
    properties:
      data:
        description: 失败时返回空
      status_code:
        description: 状态码，0-成功，其他值-失败
        type: integer
      status_msg:
        description: 返回状态描述
        type: string
    type: object
  db.GetCinemasRow:
    properties:
      avatar:
        type: string
      cols:
        type: integer
      id:
        type: integer
      name:
        type: string
      rows:
        type: integer
      total:
        type: integer
    type: object
...
  reply.CheckCinemaByName:
    properties:
      exist:
        type: boolean
    type: object
...
  reply.GetCinemaByID:
    properties:
      avatar:
        type: string
      cols:
        type: integer
      id:
        type: integer
      name:
        type: string
      rows:
        type: integer
    type: object
  reply.GetCinemas:
    properties:
      list:
        items:
          $ref: '#/definitions/db.GetCinemasRow'
        type: array
    type: object
...
  request.CreateCinema:
    properties:
      avatar:
        description: 封面链接
        type: string
      cols:
        description: 列数
        maximum: 20
        minimum: 1
        type: integer
      name:
        description: 影厅名
        maxLength: 32
        minLength: 1
        type: string
      rows:
        description: 行数
        maximum: 10
        minimum: 1
        type: integer
    required:
    - cols
    - name
    - rows
    type: object
...
  request.DeleteCinema:
    properties:
      cinema_id:
        description: 影厅ID
        minimum: 1
        type: integer
    required:
    - cinema_id
    type: object
...
  request.UpdateCinema:
    properties:
      cinema_id:
        minimum: 1
        type: integer
      new_avatar:
        description: 不填表示不更新此值
        type: string
      new_name:
        description: 不填表示不更新此值
        type: string
    required:
    - cinema_id
    type: object
...
host: ttms.humraja.xyz
info:
  contact: {}
  description: ttms影院管理系统
  license:
    name: raja,moonman
  title: ttms
  version: "1.0"
paths:
  /cinema/check_name:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 影厅名
        in: query
        maxLength: 32
        minLength: 1
        name: cinema_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.CheckCinemaByName'
              type: object
      security:
      - BasicAuth: []
      summary: 检查影厅名是否已经存在
      tags:
      - cinema
  /cinema/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: '影厅: 名字，图片，行数，列数'
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CreateCinema'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.State'
      security:
      - BasicAuth: []
      summary: 创建影厅
      tags:
      - cinema
  /cinema/delete:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 影厅id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeleteCinema'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.State'
      security:
      - BasicAuth: []
      summary: 删除影厅(不能删除存在演出计划的影厅)
      tags:
      - cinema
  /cinema/get:
    get:
      consumes:
      - application/json
      parameters:
      - description: 影厅ID
        in: query
        minimum: 1
        name: cinema_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.GetCinemaByID'
              type: object
      security:
      - BasicAuth: []
      summary: 获取影厅详细信息
      tags:
      - cinema
  /cinema/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - default: 1
        description: 第几页
        in: query
        name: page
        type: integer
      - default: 5
        description: 每页大小
        in: query
        maximum: 10
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.State'
            - properties:
                data:
                  $ref: '#/definitions/reply.GetCinemas'
              type: object
      security:
      - BasicAuth: []
      summary: 分页获取所有影厅
      tags:
      - cinema
  /cinema/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 是否存在
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UpdateCinema'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.State'
      security:
      - BasicAuth: []
      summary: 更新影厅名和图片链接
      tags:
      - cinema
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"